@@ -28,52 +28,69 @@
     "update": "ng update @nrwl/workspace",
     "workspace-schematic": "nx workspace-schematic",
     "dep-graph": "nx dep-graph",
-    "help": "nx help"
+    "help": "nx help",
+    "db:migration:generate": "node tools/migration.js"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "9.0.0",
-    "@angular/common": "9.0.0",
-    "@angular/compiler": "9.0.0",
-    "@angular/core": "9.0.0",
-    "@angular/forms": "9.0.0",
-    "@angular/platform-browser": "9.0.0",
-    "@angular/platform-browser-dynamic": "9.0.0",
-    "@angular/router": "9.0.0",
+    "@angular/animations": "9.1.0",
+    "@angular/common": "9.1.0",
+    "@angular/compiler": "9.1.0",
+    "@angular/core": "9.1.0",
+    "@angular/forms": "9.1.0",
+    "@angular/platform-browser": "9.1.0",
+    "@angular/platform-browser-dynamic": "9.1.0",
+    "@angular/router": "9.1.0",
+    "@hapi/joi": "^17.1.1",
     "@nestjs/common": "^6.8.3",
+    "@nestjs/config": "^0.4.0",
     "@nestjs/core": "^6.8.3",
     "@nestjs/platform-express": "^6.8.3",
-    "@nrwl/angular": "^9.0.1",
+    "@nestjs/swagger": "^4.5.1",
+    "@nestjs/typeorm": "^7.0.0",
+    "@nrwl/angular": "^9.1.4",
+    "chalk": "^3.0.0",
+    "class-transformer": "^0.2.3",
+    "class-validator": "^0.11.1",
     "core-js": "^3.4.2",
     "ng-zorro-antd": "^8.5.2",
     "reflect-metadata": "^0.1.13",
     "rxjs": "~6.5.3",
+    "sqlite3": "^4.1.1",
+    "typeorm": "0.2.22",
     "zone.js": "^0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.900.1",
-    "@angular/cli": "9.0.1",
-    "@angular/compiler-cli": "9.0.0",
-    "@angular/language-service": "9.0.0",
+    "@angular-devkit/build-angular": "0.901.0",
+    "@angular/cli": "9.1.0",
+    "@angular/compiler-cli": "9.1.0",
+    "@angular/language-service": "9.1.0",
     "@nestjs/schematics": "^6.7.0",
     "@nestjs/testing": "^6.8.3",
-    "@nrwl/cypress": "9.0.1",
-    "@nrwl/jest": "9.0.1",
-    "@nrwl/nest": "9.0.1",
-    "@nrwl/node": "9.0.1",
-    "@nrwl/workspace": "9.0.1",
+    "@nrwl/cypress": "9.1.4",
+    "@nrwl/jest": "9.1.4",
+    "@nrwl/nest": "9.1.4",
+    "@nrwl/node": "9.1.4",
+    "@nrwl/workspace": "9.1.4",
+    "@types/express": "^4.17.3",
+    "@types/hapi__joi": "^16.0.12",
     "@types/jest": "24.0.9",
     "@types/node": "^12.12.12",
+    "circular-dependency-plugin": "^5.2.0",
+    "clean-webpack-plugin": "^3.0.0",
     "codelyzer": "^5.2.0",
     "cypress": "^3.8.2",
-    "dotenv": "6.2.0",
-    "eslint": "6.1.0",
+    "dotenv": "^8.2.0",
+    "eslint": "6.8.0",
     "jest": "24.1.0",
     "jest-preset-angular": "8.0.0",
-    "prettier": "1.18.2",
+    "prettier": "1.19.1",
+    "prompt": "^1.0.0",
+    "swagger-ui-express": "^4.1.4",
     "ts-jest": "24.0.0",
     "ts-node": "~8.5.2",
-    "tslint": "~5.20.1",
-    "typescript": "~3.7.4"
+    "tslint": "6.0.0",
+    "tslint-config-prettier": "^1.18.0",
+    "typescript": "~3.8.3"
   }
 }
