@@ -1,4 +1,6 @@
 export const environment = {
   production: true,
-  traceRoute: false
+  traceRoute: false,
+  baseUrl: 'http://localhost:3000',
+  apiPath: 'api'
 };
