@@ -4,7 +4,9 @@
 
 export const environment = {
   production: false,
-  traceRoute: true
+  traceRoute: true,
+  baseUrl: 'http://localhost:3000',
+  apiPath: 'api'
 };
 
 /*
@@ -14,4 +16,4 @@ export const environment = {
  * This import should be commented out in production mode because it will have a negative impact
  * on performance if an error is thrown.
  */
-// import 'zone.js/dist/zone-error';  // Included with Angular CLI.
+import 'zone.js/dist/zone-error'; // Included with Angular CLI.
