@@ -1,7 +1,20 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { ModuleWithProviders, NgModule } from '@angular/core';
+import { EnvironmentConfig, ENV_CONFIG } from './environment-config.interface';
 
 @NgModule({
   imports: [CommonModule]
 })
-export class HttpModule {}
+export class HttpModule {
+  static forRoot(config: EnvironmentConfig): ModuleWithProviders<HttpModule> {
+    return {
+      ngModule: HttpModule,
+      providers: [
+        {
+          provide: ENV_CONFIG,
+          useValue: config
+        }
+      ]
+    };
+  }
+}
